+++ src/mruby_integration/models/image.cpp
@@ -106,6 +106,10 @@ void append_models_Image(mrb_state *mrb) {
         unload_image(self)
       end
 
+      def export(path)
+        export_image(self, path)
+      end
+
       def self.generate(width:, height:, colour: RAYWHITE)
         generate_image_colour(width, height, colour)
       end